requirements.txt
pygame
ibm_watson 
ibm-cloud-sdk-core
PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
fuzzywuzzy
python-Levenshtein
